factor restaurant related things into new file
